Tabla de departamentos funcional.

+++ src/main/kotlin/models/Departamento.kt
@@ -3,7 +3,7 @@ package models
 data class Departamento(
     val id: Int =  0,
     val nombre : String = "",
-    val presupuesto: Int = 0,
+    val presupuesto: Double = 0.0,
     // Donde se guardan los empleados
     //val listadoEmpleado: List<Empleado>
 )
